Converter moved to the pkg folder
