+++ internal/controller/runtime/fsm/runtime_fsm_patch_shoot.go
@@ -2,11 +2,11 @@ package fsm
 
 import (
 	"context"
+	gardener_shoot "github.com/kyma-project/infrastructure-manager/pkg/gardener/shoot"
 
 	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
 	"github.com/kyma-project/infrastructure-manager/internal/config"
-	gardener_shoot "github.com/kyma-project/infrastructure-manager/internal/gardener/shoot"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/utils/ptr"
 	ctrl "sigs.k8s.io/controller-runtime"
