+++ internal/controller/runtime/suite_test.go
@@ -19,6 +19,7 @@ package runtime
 import (
 	"context"
 	"encoding/json"
+	gardener_shoot "github.com/kyma-project/infrastructure-manager/pkg/gardener/shoot"
 	"path/filepath"
 	"testing"
 	"time"
@@ -29,7 +30,6 @@ import (
 	"github.com/kyma-project/infrastructure-manager/internal/auditlogging"
 	"github.com/kyma-project/infrastructure-manager/internal/config"
 	"github.com/kyma-project/infrastructure-manager/internal/controller/runtime/fsm"
-	gardener_shoot "github.com/kyma-project/infrastructure-manager/internal/gardener/shoot"
 	. "github.com/onsi/ginkgo/v2" //nolint:revive
 	. "github.com/onsi/gomega"    //nolint:revive
 	"github.com/pkg/errors"
