+++ pkg/gardener/shoot/extender/provider.go
@@ -1,15 +1,15 @@
 package extender
 
 import (
+	"github.com/kyma-project/infrastructure-manager/pkg/gardener/shoot/hyperscaler"
+	"github.com/kyma-project/infrastructure-manager/pkg/gardener/shoot/hyperscaler/aws"
+	"github.com/kyma-project/infrastructure-manager/pkg/gardener/shoot/hyperscaler/azure"
+	"github.com/kyma-project/infrastructure-manager/pkg/gardener/shoot/hyperscaler/gcp"
+	"github.com/kyma-project/infrastructure-manager/pkg/gardener/shoot/hyperscaler/openstack"
 	"slices"
 
 	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
-	"github.com/kyma-project/infrastructure-manager/internal/gardener/shoot/hyperscaler"
-	"github.com/kyma-project/infrastructure-manager/internal/gardener/shoot/hyperscaler/aws"
-	"github.com/kyma-project/infrastructure-manager/internal/gardener/shoot/hyperscaler/azure"
-	"github.com/kyma-project/infrastructure-manager/internal/gardener/shoot/hyperscaler/gcp"
-	"github.com/kyma-project/infrastructure-manager/internal/gardener/shoot/hyperscaler/openstack"
 	"github.com/pkg/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 )
