+++ pkg/gardener/shoot/hyperscaler/aws/config.go
