+++ pkg/gardener/shoot/hyperscaler/aws/config_test.go
