+++ pkg/gardener/shoot/hyperscaler/azure/config.go
