+++ pkg/gardener/shoot/hyperscaler/azure/config_test.go
