+++ pkg/gardener/shoot/hyperscaler/azure/model.go
