+++ pkg/gardener/shoot/hyperscaler/azure/zones.go
