+++ pkg/gardener/shoot/hyperscaler/const.go
