+++ pkg/gardener/shoot/hyperscaler/gcp/config.go
