+++ pkg/gardener/shoot/hyperscaler/gcp/config_test.go
