+++ pkg/gardener/shoot/hyperscaler/openstack/config.go
