+++ pkg/gardener/shoot/hyperscaler/openstack/config_test.go
